validation: Fix signed/unsigned comparison in txmempool.cpp
- Change totalSizeWithAncestors from size_t to int64_t
- Resolves compiler error: comparison of signed vs unsigned integers
- Aligns with Bitcoin commit 92de74ef181b42d774bc6b12329bc0c27caf0081

+++ src/txmempool.cpp
@@ -201,7 +201,7 @@ bool CTxMemPool::CalculateAncestorsAndCheckLimits(size_t entry_size,
                                                   uint64_t limitDescendantSize,
                                                   std::string &errString) const
 {
-    size_t totalSizeWithAncestors = entry_size;
+    int64_t totalSizeWithAncestors = entry_size;
 
     while (!staged_ancestors.empty()) {
         const CTxMemPoolEntry& stage = staged_ancestors.begin()->get();
